Encoder preset needs to allow more than 128 stations
Change enc_preset type to allow more than 128 stations to be allowed to be selected from the rotary encoder.

+++ Esp32_radio/Esp32_radio.ino
@@ -3983,7 +3983,7 @@ String getFSfilename ( String &nodeID )
 //**************************************************************************************************
 void chk_enc()
 {
-  static int8_t  enc_preset ;                                 // Selected preset
+  static int16_t  enc_preset ;                                 // Selected preset
   static String  enc_nodeID ;                                 // Node of selected track
   static String  enc_filename ;                               // Filename of selected track
   String         tmp ;                                        // Temporary string
